requirements.txt
pandas
duckdb
plotly
streamlit
langchain
chromadb
sentence-transformers
tavily-python
python-dotenv